Use FetchContent for Boost and fix header install
